[autoparallel] refactor the runtime apply pass and add doc string to passes
