+++ tests/test_auto_parallel/test_tensor_shard/test_resnet_block_runtime.py
@@ -17,10 +17,8 @@ from colossalai.auto_parallel.tensor_shard.solver.options import SolverOptions
 from colossalai.auto_parallel.tensor_shard.solver.solver import Solver
 from colossalai.auto_parallel.tensor_shard.solver.strategies_constructor import StrategiesConstructor
 from colossalai.device.device_mesh import DeviceMesh
-from colossalai.fx.passes.experimental.adding_shape_consistency_pass_v2 import (
-    shape_consistency_pass,
-    solution_annotatation_pass,
-)
+from colossalai.fx.passes.autoparallel.runtime_apply_pass import runtime_apply_pass
+from colossalai.fx.passes.autoparallel.runtime_preparation_pass import runtime_preparation_pass
 from colossalai.fx.tracer.tracer import ColoTracer
 from colossalai.initialize import launch
 from colossalai.logging import disable_existing_loggers
@@ -153,8 +151,8 @@ def check_apply_bottleneck(rank, world_size, port):
     print(solution)
     for index, node in enumerate(graph.nodes):
         print(node.name, node.strategies_vector[solution[index]].name)
-    sharding_spec_dict, origin_spec_dict, comm_actions_dict = solution_annotatation_pass(gm, solution, device_mesh)
-    shape_consistency_pass(gm)
+    gm, sharding_spec_dict, origin_spec_dict, comm_actions_dict = runtime_preparation_pass(gm, solution, device_mesh)
+    gm = runtime_apply_pass(gm)
     gm.recompile()
     nodes = [node for node in gm.graph.nodes]
     # TODO: wrap the gm to avoid the influence of the user training code
